// src/protocol/types.ts
export interface LoginResponse {
  Name?: string;
  BanReason?: string;
  ResultCode: E.LoginResponse_Code;
  SessionToken?: string;
}

export interface ChannelHeader {
  ChannelName?: string;
  IsDeluge: boolean;
  Password?: string;
  Topic?: Topic;
  Users?: string[];
}

export interface Topic {
  SetBy?: string;
  SetDate?: string;
  Text?: string;
}

export interface ChangeTopic {
  ChannelName?: string;
  Topic?: Topic;
}

export interface JoinChannel {
  ChannelName?: string;
  Password?: string;
}

export interface LeaveChannel {
  ChannelName?: string;
}

export interface ChannelUserAdded {
  ChannelName?: string;
  UserName?: string;
}

export interface ChannelUserRemoved {
  ChannelName?: string;
  UserName?: string;
}

export interface JoinChannelResponse {
  Channel?: ChannelHeader;
  ChannelName?: string;
  Reason?: string;
  Success: boolean;
}

export interface User {
  AccountID: number;
  Avatar?: string;
  AwaySince?: string;
  BanMute: boolean;
  BanVotes: boolean;
  BanSpecChat: boolean;
  BattleID?: number;
  Clan?: string;
  Country?: string;
  DisplayName?: string;
  Faction?: string;
  InGameSince?: string;
  IsAdmin: boolean;
  IsBot: boolean;
  LobbyVersion?: string;
  Name?: string;
  SteamID?: string;
  Badges?: string[];
  Icon?: string;
  EffectiveMmElo: number;
  EffectiveElo: number;
  Level: number;
  Rank: number;
}

export interface UserDisconnected {
  Name?: string;
  Reason?: string;
}

export interface Say {
  IsEmote: boolean;
  Place: E.SayPlace;
  Ring: boolean;
  Target?: string;
  Text?: string;
  Time?: string;
  User?: string;
  Source?: E.SaySource;
}
// ...
export interface KickFromChannel {
  ChannelName?: string;
  Reason?: string;
  UserName?: string;
}

export interface ForceJoinChannel {
  ChannelName?: string;
  UserName?: string;
}