// src/protocol/enums.ts
export enum SayPlace {
  Channel = 0,
  Battle = 1,
  User = 2,
  BattlePrivate = 3,
  Game = 4,
  MessageBox = 5,
}